building cart

+++ Models/CartDetail.cs
@@ -9,7 +9,7 @@ namespace WebApp.Models
         public int ShoppingCartId { get; set; }
         public int ProductId { get; set; }
         public int Quantity { get; set; }
-        public double UnitPriced { get; set; }
+        public int UnitPriced { get; set; }
         public Product Product { get; set; }
         public ShoppingCart ShoppingCart { get; set; }
     }
